refactor: analysis service interface 생성
